chore(build-push.yml): add GitHub Actions workflow to build and push Docker image
This commit adds a new file `.github/workflows/build-push.yml` which contains a GitHub Actions workflow. The workflow is triggered on a push event to the `master` branch.

The workflow consists of a single job `build-and-push` which runs on `ubuntu-latest` environment.

The steps in the job are as follows:
1. Checkout Repository: This step uses the `actions/checkout@v2` action to checkout the repository.
2. Create image tags: This step uses the `docker/metadata-action@v4` action to create tags for the Docker image. The image is tagged with the repository `ghcr.io/userofficeproject/ror-gateway` and the tag is generated based on the commit SHA.
3. Build and push: This step uses the `docker/build-push-action@v3` action to build and push the Docker image. The image is built for both `linux/amd64` and `linux/arm64/v8` platforms, and the tags generated in the previous step are used for pushing the image.

The purpose of this commit is to automate the process of building and pushing the Docker image whenever there is a push to the `master` branch. This will ensure that the latest version of the image is always available for deployment.